Fix format specifier warning
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@36498 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/c64/cart/comal80.c
@@ -109,7 +109,7 @@ static uint8_t comal80_io1_peek(uint16_t addr)
 static int comal80_dump(void)
 {
     mon_out("extra eprom is installed: %s\n", extrarom ? "yes" : "no");
-    mon_out("register value: $%02x\n", currregval);
+    mon_out("register value: $%02x\n", (unsigned int)currregval);
     mon_out(" bank: %d/%d\n", currregval & 7, extrarom ? 8 : 4);
     return 0;
 }
